requirements.txt
attrs==25.1.0
click==8.1.8
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
KDEpy==1.1.11
nltk==3.9.1
numpy==2.2.3
openpyxl==3.1.5
pandas==2.2.3
playwright==1.51.0
python-dateutil==2.9.0.post0
pytz==2025.1
referencing==0.36.2
regex==2024.11.6
rpds-py==0.23.1
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
xlsxwriter==3.2.9